refactor: extract super ops control center
